Requirement: R1871
1. Fixed a couple of issues with IEntityBroker and EntityBroker.  The Count method now has a return type of int rather than long, because that is what Hibernate supports.
2. Fixed an issue with AuthorityTokenBroker.

+++ Enterprise/Authentication/AuthenticationService.cs
@@ -5,6 +5,7 @@ using ClearCanvas.Enterprise.Core;
 using ClearCanvas.Enterprise.Authentication.Brokers;
 using ClearCanvas.Common.Utilities;
 using ClearCanvas.Common;
+using ClearCanvas.Enterprise.Common;
 
 namespace ClearCanvas.Enterprise.Authentication
 {
@@ -23,7 +24,7 @@ namespace ClearCanvas.Enterprise.Authentication
             UserSearchCriteria where = new UserSearchCriteria();
             where.UserName.EqualTo(userName);
 
-            long count = PersistenceContext.GetBroker<IUserBroker>().Count(where);
+            int count = PersistenceContext.GetBroker<IUserBroker>().Count(where);
             return count == 1;
         }
